// resources/js/api/generated/models/buildCacheInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export interface BuildCacheInfo {
  created: string;
  description: string;
  id: string;
  in_use: boolean;
  /** @nullable */
  last_used: string | null;
  removal: string;
  shared: boolean;
  size: number;
  type: string;
  usage_count: number;
}