// resources/js/api/generated/models/maintenanceInfo.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { BuildCacheSummary } from './buildCacheSummary';
import type { ContainerSummary } from './containerSummary';
import type { ImageSummary } from './imageSummary';
import type { NetworkSummary } from './networkSummary';
import type { SystemInfo } from './systemInfo';
import type { VolumeSummary } from './volumeSummary';

export interface MaintenanceInfo {
  build_cache_summary: BuildCacheSummary;
  container_summary: ContainerSummary;
  image_summary: ImageSummary;
  last_updated: string;
  network_summary: NetworkSummary;
  system_cleanup_covers: string[];
  system_info: SystemInfo;
  volume_summary: VolumeSummary;
}